backend/requirements.txt
fastapi
uvicorn
sqlalchemy
apscheduler
email-validator
python-dotenvcd